pyproject.toml
[tool.poetry]
name = "washday"
version = "0.1.0"
description = ""
authors = ["r0w4nr00t <r0w4nr00t@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-filter = "^23.5"
pyyaml = "^6.0.1"
django-phonenumber-field = "^7.3.0"
phonenumbers = "^8.13.30"
django-split-settings = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
